Refactor layout, remove complex client-side logic, and simplify app structure
