+++ package.json
@@ -14,14 +14,14 @@
     "@supabase/supabase-js": "^2.49.10",
     "@vercel/analytics": "^1.5.0",
     "@zanixongroup/uploader": "^2.0.7",
+    "critters": "^0.0.23",
     "framer-motion": "^12.16.0",
     "next": "15.3.3",
     "next-pwa": "^5.6.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hot-toast": "^2.5.2",
     "react-icons": "^5.5.0",
-    "react-image-lightbox": "^5.1.4",
     "react-intersection-observer": "^9.16.0",
     "react-loading-skeleton": "^3.5.0",
     "react-masonry-css": "^1.0.16",
